+++ src/index.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './store/store-index';
-import { loadProductsFromMongo } from './actions/actions-data';
-//import { getProducts } from './actions/actions-products';
+import { loadDataForAdmin } from './actions/actions-data';
 import RoutesDefinition from './routes';
 import './App.css';
 
@@ -19,6 +18,6 @@ render(
 document.getElementById('root')
 );
 
-//store.dispatch(loadProductsFromMongo())
+store.dispatch(loadDataForAdmin())