+++ src/reducers/reducer-products.js
@@ -29,12 +29,14 @@ const productsReducer =  function (state = initialState, action) {
             const filterColor = state.selectedCategory !== null  ? "category" : "color";
             return  {...state, visibleProducts: state.selectedColor, filter: filterColor};
         case REMOVE_CATEGORY:
-            state.selectedCategory = null;
-            state.selectedColor = null;
+            //state.selectedCategory = null;
+            //state.selectedColor = null;
+            state.selectedCategory !== null ?  state.selectedCategory = null : state;
             return  {...state, visibleProducts: state.products, filter: ""};
         case REMOVE_COLOR:
-            state.selectedColor = null;
-            state.selectedCategory = null;
+            //state.selectedColor = null;
+            //state.selectedCategory = null;
+            state.selectedColor !== null ?  state.selectedColor = null : state;
             return  {...state, visibleProducts: state.products, filter: ""};
         default:
             return state;