+++ changelog.md
@@ -30,6 +30,7 @@ TROUBLES
 - 2.06.19
     - próby redux/router - strona produktu
     - panel admin wyświetlanie listy dostępnych produktów i ich stanów magazynowych - bez stylowania
+    - oodzielny reducer dla admin
 - 1.06.19
     - rozwiązanie problemów z reduxem - były źle czyszczone filtry
 - 25.05.19